+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- added : stat popup shown from table column tooltip btn
 - changed : description info with correct padding and width on mobile
 
 ## 0.9.4 (2024-11-12)